Add some support for icarus verilog simulation
- refactor the octavo file lists into Misc.octavo_files
- generate script for iverilog

Note that in order to generate a vcd you should run vvp with

```
$ vvp -s a.out
vvp> $dumpvars
vvp> cont
...........
vvp> finish
```

we could maybe do this with a script.  I have also temporarily reduced
the cycle count of the simulation while I debug the code.  It does not
match between iverilog and modelsim at the moment.